cart step extras now load outside initial runtime
- moves promo code form runtime behind the coupon checkbox interaction

- extracts transport/payment validation from page utilities

- keeps checkout validation tests with the extracted utility
